+++ TheStageCore.xcframework/Info.plist
@@ -6,20 +6,6 @@
 	<array>
 		<dict>
 			<key>BinaryPath</key>
-			<string>TheStageCore.framework/Versions/A/TheStageCore</string>
-			<key>LibraryIdentifier</key>
-			<string>macos-arm64</string>
-			<key>LibraryPath</key>
-			<string>TheStageCore.framework</string>
-			<key>SupportedArchitectures</key>
-			<array>
-				<string>arm64</string>
-			</array>
-			<key>SupportedPlatform</key>
-			<string>macos</string>
-		</dict>
-		<dict>
-			<key>BinaryPath</key>
 			<string>TheStageCore.framework/TheStageCore</string>
 			<key>LibraryIdentifier</key>
 			<string>ios-arm64</string>
@@ -31,6 +17,20 @@
 			</array>
 			<key>SupportedPlatform</key>
 			<string>ios</string>
+		</dict>
+		<dict>
+			<key>BinaryPath</key>
+			<string>TheStageCore.framework/Versions/A/TheStageCore</string>
+			<key>LibraryIdentifier</key>
+			<string>macos-arm64</string>
+			<key>LibraryPath</key>
+			<string>TheStageCore.framework</string>
+			<key>SupportedArchitectures</key>
+			<array>
+				<string>arm64</string>
+			</array>
+			<key>SupportedPlatform</key>
+			<string>macos</string>
 		</dict>
 	</array>
 	<key>CFBundlePackageType</key>
